Java Generic Exercises

+++ src/generic/ex/ex03/NumberDataProcessor.java
@@ -1,8 +1,9 @@
 package generic.ex.ex03;
 
 public class NumberDataProcessor implements DataProcessor<Integer> {
+
     @Override
     public Integer processData(Integer integer) {
-        return integer;
+        return integer * 2;
     }
 }
